requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Computer Vision
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
pandas>=2.0.0

# Configuration
pyyaml>=6.0

# Progress Bars
tqdm>=4.65.0

# Logging
tensorboard>=2.13.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

setuptools>=67.0.0
kaggle>=1.6.6
omegaconf>=2.1.0